Using size_t instead of unsigned in a macro

+++ include-blang/annotations/annotations.h
@@ -119,11 +119,11 @@ _FUNCTION_FROM_POINTER_TO_VOID(writes_to)
 bool __atomic_has_taken_value_local(__local unsigned *atomic_array, size_t offset, unsigned value);
 bool __atomic_has_taken_value_global(__global unsigned *atomic_array, size_t offset, unsigned value);
 
-_CLC_OVERLOAD _CLC_INLINE bool __atomic_has_taken_value(__local unsigned *atomic_array, unsigned offset, unsigned value) {
+_CLC_OVERLOAD _CLC_INLINE bool __atomic_has_taken_value(__local unsigned *atomic_array, size_t offset, unsigned value) {
     return __atomic_has_taken_value_local(atomic_array, offset, value);
 }
 
-_CLC_OVERLOAD _CLC_INLINE bool __atomic_has_taken_value(__global unsigned *atomic_array, unsigned offset, unsigned value) {
+_CLC_OVERLOAD _CLC_INLINE bool __atomic_has_taken_value(__global unsigned *atomic_array, size_t offset, unsigned value) {
     return __atomic_has_taken_value_global(atomic_array, offset, value);
 }
 #endif
